Move executable property logic to TemplatePlugin
